# specs/001-create-a-rest/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Test Results Management API
  description: |
    REST API for CRUD operations on end-to-end test execution results.
    Primary integration via GitHub Actions CI/CD pipelines using bearer token authentication.

    ## Authentication
    All endpoints require JWT bearer token authentication. Tokens are obtained through
    GitHub SSO for human users or generated for automated systems (CI/CD pipelines).

    ## Integration Patterns
    - **GitHub Actions**: Use bearer tokens stored in repository/organization secrets
    - **CI/CD Systems**: Automated test result submission with artifact upload
    - **Development**: Local testing with personal access tokens
  version: 0.4.0
  contact:
    name: Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.testresults.dev
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TestFramework:
      type: object
      required: [name, version]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          pattern: '^[a-z0-9-]+$'
          maxLength: 50
          example: "playwright"
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          example: "1.40.0"
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    TestEnvironment:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 100
          example: "staging"
        browser:
          type: string
          enum: [chromium, firefox, webkit, chrome, edge]
          example: "chromium"
        os:
          type: string
          enum: [ubuntu-22.04, ubuntu-20.04, macos-13, macos-12, windows-2022, windows-2019]
          example: "ubuntu-22.04"
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "RESOURCE_NOT_FOUND"
        field:
          type: string
          example: "suite_id"
paths:
...
  /frameworks:
    get:
      summary: List test frameworks
      tags: [Frameworks]
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Filter by framework name
      responses:
        '200':
          description: List of test frameworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestFramework'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create test framework
      tags: [Frameworks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestFramework'
      responses:
        '201':
          description: Framework created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestFramework'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Framework already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /frameworks/{framework_id}:
    get:
      summary: Get test framework
      tags: [Frameworks]
      parameters:
        - name: framework_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Framework details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestFramework'
        '404':
          description: Framework not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update test framework
      tags: [Frameworks]
      parameters:
        - name: framework_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestFramework'
      responses:
        '200':
          description: Framework updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestFramework'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Framework not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete test framework
      tags: [Frameworks]
      parameters:
        - name: framework_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Framework deleted successfully
        '404':
          description: Framework not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot delete framework with existing suites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Test Environments
  /environments:
    get:
      summary: List test environments
      tags: [Environments]
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Filter by environment name
        - name: browser
          in: query
          schema:
            type: string
          description: Filter by browser type
      responses:
        '200':
          description: List of test environments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestEnvironment'

    post:
      summary: Create test environment
      tags: [Environments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestEnvironment'
      responses:
        '201':
          description: Environment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestEnvironment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /environments/{environment_id}:
    get:
      summary: Get test environment
      tags: [Environments]
      parameters:
        - name: environment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Environment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestEnvironment'
        '404':
          description: Environment not found

    put:
      summary: Update test environment
      tags: [Environments]
      parameters:
        - name: environment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestEnvironment'
      responses:
        '200':
          description: Environment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestEnvironment'

    delete:
      summary: Delete test environment
      tags: [Environments]
      parameters:
        - name: environment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Environment deleted successfully
        '409':
          description: Cannot delete environment with existing suites

  # Test Suites
...